fix(arch): implement first P0 batch (user-approved) — K27 graph_conflict +1 overflow + lv_SAFE_ADD convergence to lv_safe_add_int (new wrapper), K28 depth_limits.h single authority table (128 x4 collapsed + recursion/runtime_guard/config reference it) + parser_safety.h missing declarations, K66 4 compile-time pool-size _Static_asserts (GeomNode/Constraint/SymbolicCoord/ProofStep), K67 func_block instantiate 6-arg out-param fix (was 5-arg TypeError swallowed by safe_instantiate), E11 TEST_ASSERT_EQ/DOUBLE forward to lv_ASSERT_* with swapped args (old signature kept), I3 lv_ensure_capacity64 size_t authority + sparse_linear_algebra migration; local build all + ctest 288/288 pass; register batch 153

+++ core/src/layer3_geometry/constraint_graph/graph_conflict.c
@@ -609,7 +609,9 @@ int **graph_detect_conflicts(const ConstraintGraph *graph, int *out_conflict_cou
     }
 
     /* adj: node_id -> 约束索引的扁平数组 */
-    size_t adj_total = (size_t) (max_node_id + 1) * lv_ADJ_MAX_PER_NODE;
+    /* K27：先提升 size_t 再 +1——原 (size_t)(max_node_id + 1) 在 int 域先加，
+     * max_node_id == INT_MAX 时为有符号溢出 UB */
+    size_t adj_total = ((size_t) max_node_id + 1) * (size_t) lv_ADJ_MAX_PER_NODE;
     if (adj_total > (size_t) INT_MAX) {
         lv_free((void **) &conflicts);
         lv_free((void **) out_conflict_sizes);
